ManipulatorInterfejs Buttons on main window WIP     to do:     -trasparecy     -proximiti     -schortcuts
