+++ manipulator/AbstractManipulatroInterfejs.py
@@ -1,0 +1,40 @@
+from PyQt5.QtWidgets import QWidget, QGridLayout, QPushButton, QAction, QHBoxLayout
+from PyQt5.QtCore import Qt
+
+
+class _AbstractManipulatorInterfejs(QWidget):
+
+    def __init__(self, ManipulatorObject, *args, **kwargs):
+        super(_AbstractManipulatorInterfejs, self).__init__( *args, **kwargs)
+
+        self.Manipulator = ManipulatorObject
+
+        self._layout = QGridLayout()
+
+        self.buttonsNames = ['/\\', "<", ">", '\/']
+
+        self.buttons = [QPushButton(name) for name in self.buttonsNames]
+
+        [self._layout.addWidget(value, i, j) for j, i, value in zip([2, 3, 3, 4], [3, 2, 4, 3], self.buttons)]
+
+        self.actions = [QAction("&dw", self), QAction("&lf", self), QAction("&ri", self), QAction("&up", self)]
+
+        self.fun = [self._key_down, self._key_left, self._key_right, self._key_up]
+
+        [a.triggered.connect(f) for a, f in zip(self.actions, self.fun)]
+
+        [button.released.connect(f) for f, button in zip(self.fun, self.buttons)]
+
+        self.setLayout(self._layout)
+
+    def _key_up(self):
+        self.Manipulator.up()
+
+    def _key_left(self):
+        self.Manipulator.left()
+
+    def _key_right(self):
+        self.Manipulator.right()
+
+    def _key_down(self):
+        self.Manipulator.down()
